Блок-схема скрипта
